Refactored to a separate NotifyingHttpClient class
